+++ include/kautham/problem/link.h
@@ -88,7 +88,7 @@ namespace Kautham{
 	  *		associated link and put this link in the origin of absolute frame. 
       *		You can build a complete robot, if you adding progresively a each Link
 	  *		from absolute coordinates frame to final effector frame.*/
-      Link(string ivFile, string collision_ivFile, float scale,
+      Link(string ivFile, string collision_ivFile, double scale,
            APPROACH Type, bool useBBOX = false);
 
       //!	Constructor.
@@ -96,7 +96,7 @@ namespace Kautham{
       *		associated link and put this link in the origin of absolute frame.
       *		You can build a complete robot, if you adding progresively a each Link
       *		from absolute coordinates frame to final effector frame.*/
-      Link(SoSeparator *visual_model, SoSeparator *collision_model, float scale,
+      Link(SoSeparator *visual_model, SoSeparator *collision_model, double scale,
            APPROACH Type, bool useBBOX = false);
 
 	  //! Function to set \f$ \alpha \f$ parameter.
